# http/swagger.json
{
    "schemes": [
        "http"
    ],
    "swagger": "2.0",
    "info": {
        "description": "Centrifuge OS Node API",
        "title": "Centrifuge OS Node API",
        "contact": {
            "name": "Centrifuge",
            "url": "https://github.com/centrifuge/pod",
            "email": "hello@centrifuge.io"
        },
        "license": {
            "name": "MIT"
        },
        "version": "3.0.0"
    },
    "host": "localhost:8082",
    "basePath": "/",
    "paths": {
...
        "/v2/accounts": {
            "get": {
                "description": "Returns all the accounts in the node.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Accounts"
                ],
                "summary": "Returns all the accounts in the node.",
                "operationId": "get_accounts_v2",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/coreapi.Accounts"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/httputils.HTTPError"
                        }
                    }
                }
            }
        },
        "/v2/accounts/generate": {
            "post": {
                "description": "Generates a new account with defaults.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Accounts"
                ],
                "summary": "Generates a new account with defaults.",
                "operationId": "generate_account_v2",
                "parameters": [
                    {
                        "description": "Generate Account Payload",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/coreapi.GenerateAccountPayload"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/coreapi.Account"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/httputils.HTTPError"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/httputils.HTTPError"
                        }
                    }
                }
            }
        },
        "/v2/accounts/self": {
            "get": {
                "description": "Returns the account associated with the identity provided in the JW3T auth token.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Accounts"
                ],
                "summary": "Returns the account associated with the identity provided in the JW3T auth token.",
                "operationId": "get_self_v2",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/coreapi.Account"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/httputils.HTTPError"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/httputils.HTTPError"
                        }
                    }
                }
            }
        },
        "/v2/accounts/{account_id}": {
            "get": {
                "description": "Returns the account associated with accountID.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Accounts"
                ],
                "summary": "Returns the account associated with accountID.",
                "operationId": "get_account_v2",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Account ID",
                        "name": "account_id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/coreapi.Account"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/httputils.HTTPError"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/httputils.HTTPError"
                        }
                    }
                }
            }
        },
        "/v2/accounts/{account_id}/sign": {
            "post": {
                "description": "Signs and returns the signature of the Payload.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Accounts"
                ],
                "summary": "Signs and returns the signature of the Payload.",
                "operationId": "account_sign_v2",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Account ID",
                        "name": "account_id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Sign request",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/coreapi.SignRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/coreapi.SignResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/httputils.HTTPError"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/httputils.HTTPError"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "coreapi.Account": {
            "type": "object",
            "properties": {
                "document_signing_public_key": {
                    "type": "array",
                    "items": {
                        "type": "integer"
                    }
                },
                "identity": {
                    "type": "string"
                },
                "p2p_public_signing_key": {
                    "type": "array",
                    "items": {
                        "type": "integer"
                    }
                },
                "pod_operator_account_id": {
                    "type": "array",
                    "items": {
                        "type": "integer"
                    }
                },
                "precommit_enabled": {
                    "type": "boolean"
                },
                "webhook_url": {
                    "type": "string"
                }
            }
        },
        "coreapi.Accounts": {
            "type": "object",
            "properties": {
                "data": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/coreapi.Account"
                    }
                }
            }
        },
...
        "coreapi.GenerateAccountPayload": {
            "type": "object",
            "properties": {
                "account": {
                    "$ref": "#/definitions/coreapi.Account"
                }
            }
        },
...
        "coreapi.SignRequest": {
            "type": "object",
            "properties": {
                "payload": {
                    "type": "string"
                }
            }
        },
        "coreapi.SignResponse": {
            "type": "object",
            "properties": {
                "payload": {
                    "type": "string"
                },
                "public_key": {
                    "type": "string"
                },
                "signature": {
                    "type": "string"
                },
                "signer_id": {
                    "type": "string"
                }
            }
        },
...
        "httputils.HTTPError": {
            "type": "object",
            "properties": {
                "message": {
                    "type": "string"
                }
            }
        },
...
    }
}